Add Ghost.cs and better win detection
Player only wins if they are in the center area of the cell now.
Ghosts are floating white cubes that hone in on the player. Lots of
values can be played with to make them easy or impossible to avoid.
Next commit will add translucency to ghosts, make player lose game
if they come into contact with a ghost too many times, and maybe
some other stuff.